fix type

+++ app/models/post.rb
@@ -3,6 +3,6 @@ class Post
   field :user, type: String
   field :story, type: String
   field :about_who, type: String
-  field :lat, type: Double
-  field :lon, type: Double
+  field :lat, type: Integer
+  field :lon, type: Integer
 end
